+++ src/services/weather-stack.service.ts
@@ -4,7 +4,7 @@ import { WsApiResponse } from '../types/ws-api.type';
 
 export class WeatherStackService {
 
-    public async get(city: string, state: string, zipCode: string): Promise<WsApiResponse> {
+    public async get(city: string, state: string, zipCode: number): Promise<WsApiResponse> {
         try {
             const response = await fetch(`http://api.weatherstack.com/current?access_key=${config?.wsApiAccessKey}&query=${city},${state},${zipCode}`);
 
@@ -14,7 +14,13 @@ export class WeatherStackService {
 
             logger.info(`Successfully received weather information for ${city}, ${state}, ${zipCode}.`);
 
-            return await response.json() as WsApiResponse;
+            const wsApiResponse = await response.json() as WsApiResponse;
+
+            if (wsApiResponse.location.country !== 'United States of America') {
+                throw new Error(`Weather information was found for ${wsApiResponse.location.country} instead of US!`);
+            }
+
+            return wsApiResponse;
         } catch (err: any) {
             logger.error(`Error fetching data: ${err.message}`);