+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-03-10
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `constants` file.
+
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): Validation for state and country.
+- [Marek Kurańda](https://github.com/mjkuranda): Schema for properties - `zipCode` length equals 5 digits.
+
 ## [0.4.0] - 2025-03-08
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): GraphQL management and `resolver` and `schema`.